Add double support for sensor interface
Add support for reading sensors that produce a double
as the value. If there is a scale, assume int64_t, else
assume double.

+++ dbus/util.hpp
@@ -5,7 +5,7 @@
 struct SensorProperties
 {
     int64_t scale;
-    int64_t value;
+    double value;
     std::string unit;
 };
